--- pycqed/instrument_drivers/physical_instruments/ZurichInstruments/wouter/zi_parameter_files/d_node_pars.txt
@@ -1,239 +1,0 @@
-[  
-  [
-    "system/fpgarevision",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/program",
-    "vector_g"
-  ],
-  [
-    "awgs/0/sweep/awgtrigs/3",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/memoryusage",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/continue",
-    "float"
-  ],
-  [
-    "awgs/0/sweep/awgtrigs/2",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/assembly",
-    "vector_g"
-  ],
-  [
-    "awgs/0/sweep/awgtrigs/1",
-    "float"
-  ],
-  [
-    "quex/sl/statemap",
-    "vector_gs"
-  ],
-  [
-    "awgs/0/sweep/awgtrigs/0",
-    "float"
-  ],
-  [
-    "awgs/0/elf/name",
-    "vector_g"
-  ],
-  [
-    "awgs/0/dio/data",
-    "vector_g"
-  ],
-  [
-    "awgs/0/sequencer/next",
-    "float"
-  ],
-  [
-    "awgs/0/elf/checksum",
-    "float"
-  ],
-  [
-    "awgs/0/waveform/data",
-    "vector_gs"
-  ],
-  [
-    "awgs/0/waveform/memoryusage",
-    "float"
-  ],
-  [
-    "awgs/0/elf/memoryusage",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/status",
-    "float"
-  ],
-  [
-    "awgs/0/sequencer/pc",
-    "float"
-  ],
-  [
-    "awgs/0/elf/data",
-    "vector_gs"
-  ],
-  [
-    "awgs/0/waveform/descriptors",
-    "vector_g"
-  ],
-  [
-    "sigins/1/max",
-    "float"
-  ],
-  [
-    "sigins/0/max",
-    "float"
-  ],
-  [
-    "sigins/0/min",
-    "float"
-  ],
-  [
-    "sigins/0/autorange",
-    "float"
-  ],
-  [
-    "sigins/1/min",
-    "float"
-  ],
-  [
-    "sigins/1/autorange",
-    "float"
-  ],
-  [
-    "sigouts/0/over",
-    "float"
-  ],
-  [
-    "sigouts/1/over",
-    "float"
-  ],
-  [
-    "quex/iavg/data/0",
-    "vector_g"
-  ],
-  [
-    "quex/iavg/data/1",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/0",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/1",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/2",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/3",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/4",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/5",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/6",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/7",
-    "vector_g"
-  ],
-  [
-    "quex/rl/data/8",
-    "vector_g"
-  ],
-  [
-    "quex/wint/weights/0/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/1/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/2/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/3/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/4/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/5/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/6/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/7/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/8/real",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/0/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/1/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/2/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/3/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/4/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/5/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/6/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/7/imag",
-    "vector_gs"
-  ],
-  [
-    "quex/wint/weights/8/imag",
-    "vector_gs"
-  ],
-
-  [
-    "dios/0/input",
-    "vector_s"
-  ]
-]
